Update MReader.lua
One the website there are two listed areas of chapters. the first, and the one used until now, pulled from the most recent 50 chapters. What this update does is instead pulls from the all chapters listing. that way of any decent length you do not have to download it 2 separate times and move the chapters from the N/A folder.


There may be issues. for example for some reason i dont see the module in the drop down anymore and dont know how to fix.